requirements.txt
django==5.1.1
django-rest-framework
gunicorn
whitenoise
django-unfold